Move svc_list to core
